until create reply system

+++ restapi/models/UserModel.py
@@ -1,8 +1,8 @@
-from sqlalchemy import Table, Column, Integer, String, DateTime, func
+from sqlalchemy import Table, Column, BigInteger, String, DateTime, func
 from config import metadata
 
 user = Table('users', metadata,
-    Column('id', Integer, primary_key=True),
+    Column('id', BigInteger, primary_key=True),
     Column('username', String(100), nullable=False),
     Column('email', String(100), unique=True, index=True, nullable=False),
     Column('password', String(100), nullable=True),
